refactor: make URLPipeLoader private
Move the URLPipeLoader class into an anonymous namespace in
electron_url_loader_factory.cc. Manual backport of 1b5db99.
