+++ filenames.gni
@@ -456,8 +456,6 @@ filenames = {
     "shell/browser/net/system_network_context_manager.h",
     "shell/browser/net/url_loader_network_observer.cc",
     "shell/browser/net/url_loader_network_observer.h",
-    "shell/browser/net/url_pipe_loader.cc",
-    "shell/browser/net/url_pipe_loader.h",
     "shell/browser/net/web_request_api_interface.h",
     "shell/browser/network_hints_handler_impl.cc",
     "shell/browser/network_hints_handler_impl.h",
